PLANET-8055 Write an e2e test for the Posts List block (grid layout)
This commit also refactors the tests for this specific block to avoid code duplication
